+++ src/TreeSitter.jl
@@ -2,6 +2,7 @@ module TreeSitter
 
 export Parser, Tree, Node, Language, Query
 export parse, traverse, children, named_children, @query_cmd
+export list_parsers
 
 include("api.jl")
 include("interface.jl")